Rename Tool_Lists to Tool_Lists.txt
convention? .txt, doesn't need to or have to be however my other files created by Deploy_Setup.sh have .txt, so for some consistency's sake, this file gets the .txt as it houses variables to be sourced by scripts, not the env. might be a better file suffix for this. If your are reading this and know of one, please let me know!